<Changed flow to only test connection per room action>
