+++ app/controllers/application_controller.rb
@@ -50,8 +50,26 @@ class ApplicationController < ActionController::Base
     end
   end
 
+  def bbb_endpoint
+    Rails.configuration.bigbluebutton_endpoint
+  end
+
+  def bbb_secret
+    Rails.configuration.bigbluebutton_secret
+  end
+
+  def remove_slash(s)
+    s.nil? ? nil : s.chomp("/")
+  end
+
   def bbb_connected?
-    Rails.configuration.bbb_connect
+    unless Rails.configuration.loadbalanced_configuration
+      BigBlueButton::BigBlueButtonApi.new(remove_slash(bbb_endpoint), bbb_secret, "0.8").test_connection
+    end
+    true
+  rescue BigBlueButton::BigBlueButtonException => exc
+    puts "ERROR: " + exc.to_s
+    false
   end
   helper_method :bbb_connected?
